Move test for #52
