+++ test/NetTopologySuite.IO.GeoJSON.Test/GeoJsonSerializerTest.cs
@@ -190,79 +190,6 @@ namespace NetTopologySuite.IO.GeoJSON.Test
             Assert.AreEqual(1, features.Count);
         }
 
-        [TestCase(RingOrientationOption.DoNotModify, "POLYGON((0 0, 10 0, 0 10, 0 0), (1 1, 8 1, 1 8, 1 1))")]
-        [TestCase(RingOrientationOption.DoNotModify, "POLYGON((0 0, 0 10, 10 0, 0 0), (1 1, 8 1, 1 8, 1 1))")]
-        [TestCase(RingOrientationOption.DoNotModify, "POLYGON((0 0, 0 10, 10 0, 0 0), (1 1, 1 8, 8 1, 1 1))")]
-        [TestCase(RingOrientationOption.EnforceRfc9746, "POLYGON((0 0, 10 0, 0 10, 0 0), (1 1, 8 1, 1 8, 1 1))")]
-        [TestCase(RingOrientationOption.EnforceRfc9746, "POLYGON((0 0, 0 10, 10 0, 0 0), (1 1, 8 1, 1 8, 1 1))")]
-        [TestCase(RingOrientationOption.EnforceRfc9746, "POLYGON((0 0, 0 10, 10 0, 0 0), (1 1, 1 8, 8 1, 1 1))")]
-        [TestCase(RingOrientationOption.NtsGeoJsonV2, "POLYGON((0 0, 10 0, 0 10, 0 0), (1 1, 8 1, 1 8, 1 1))")]
-        [TestCase(RingOrientationOption.NtsGeoJsonV2, "POLYGON((0 0, 0 10, 10 0, 0 0), (1 1, 8 1, 1 8, 1 1))")]
-        [TestCase(RingOrientationOption.NtsGeoJsonV2, "POLYGON((0 0, 0 10, 10 0, 0 0), (1 1, 1 8, 8 1, 1 1))")]
-        public void TestIssue52(RingOrientationOption ero, string wkt)
-        {
-            var rdr = new WKTReader();
-            var poly1 = (Polygon)rdr.Read(wkt);
-            var js = new JsonSerializerSettings();
-            var s = GeoJsonSerializer.Create(js, poly1.Factory, 2, ero);
-            var buffer = new StringBuilder();
-            using var w = new JsonTextWriter(new StringWriter(buffer));
-            s.Serialize(w, poly1);
-            using var r = new JsonTextReader(new StringReader(buffer.ToString()));
-            var poly2 = s.Deserialize<Polygon>(r);
-
-            switch(ero) {
-                case RingOrientationOption.DoNotModify:
-                    CheckRingOrientation(poly2.ExteriorRing, GetRingOrientation(poly1.ExteriorRing));
-                    for (int i = 0; i < poly2.NumInteriorRings; i++)
-                        CheckRingOrientation(poly2.GetInteriorRingN(i), GetRingOrientation(poly1.GetInteriorRingN(i)));
-                    break;
-
-                case RingOrientationOption.EnforceRfc9746:
-                    CheckRingOrientation(poly2.ExteriorRing, OrientationIndex.CounterClockwise);
-                    for (int i = 0; i < poly2.NumInteriorRings; i++)
-                        CheckRingOrientation(poly2.GetInteriorRingN(i), OrientationIndex.Clockwise);
-                    break;
-
-                case RingOrientationOption.NtsGeoJsonV2:
-                    CheckRingOrientation(poly2.ExteriorRing, OrientationIndex.Clockwise);
-                    for (int i = 0; i < poly2.NumInteriorRings; i++)
-                        CheckRingOrientation(poly2.GetInteriorRingN(i), OrientationIndex.CounterClockwise);
-                    break;
-            }
-
-
-        }
-
-        private static void CheckRingOrientation(LineString ring, OrientationIndex ori)
-        {
-            Assert.That(GetRingOrientation(ring), Is.EqualTo(ori));
-        }
-
-
-        private static OrientationIndex GetRingOrientation(LineString ring)
-        {
-            if (!(ring is LinearRing lr))
-                throw new ArgumentException(nameof(ring));
-
-            return lr.IsCCW ? OrientationIndex.CounterClockwise : OrientationIndex.Clockwise;
-        }
-
-        private void TestIssue52_2()
-        {
-            throw new NotImplementedException();
-        }
-
-        private void TestIssue52_1()
-        {
-            throw new NotImplementedException();
-        }
-
-        private void TestIssue52_0()
-        {
-            throw new NotImplementedException();
-        }
-
         public static MemoryStream Serialize(JsonSerializer serializer, object item)
         {
             var res = new MemoryStream();
